#1070 Finishing renames, fixed tests.
