+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/Introductions/Events/Invoke.t.cs
@@ -1,0 +1,34 @@
+[Introduction]
+    internal class TargetClass
+    {
+      private static readonly global::Metalama.Framework.RunTime.ActionEventBrokerCallbacks<global::System.EventHandler, (global::System.Object? , global::System.EventArgs)> EventFromAccessorsBrokerCallbacks_0 = new global::Metalama.Framework.RunTime.ActionEventBrokerCallbacks<global::System.EventHandler, (global::System.Object? , global::System.EventArgs)>(static (handler, me, args) => ((global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Introductions.Events.Invoke.TargetClass)me).EventFromAccessors_Invoke_Introduction(handler, args), static b => (sender, e) => b.Invoke((sender, e)), static (handler, me) => ((global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Introductions.Events.Invoke.TargetClass)me).EventFromAccessors_Introduction += handler, static (handler, me) => ((global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Introductions.Events.Invoke.TargetClass)me).EventFromAccessors_Introduction -= handler);
+      private event global::System.EventHandler EventFromAccessors_Introduction
+      {
+        add
+        {
+          global::System.Console.WriteLine("Add");
+        }
+        remove
+        {
+          global::System.Console.WriteLine("Remove");
+        }
+      }
+      private void EventFromAccessors_Invoke_Introduction(global::System.EventHandler handler, (global::System.Object? sender, global::System.EventArgs e) args)
+      {
+        global::System.Console.WriteLine("Invoke");
+        handler.Invoke(args.sender, args.e);
+      }
+      private volatile global::Metalama.Framework.RunTime.ActionEventBroker<global::System.EventHandler, (global::System.Object? , global::System.EventArgs)>? _eventFromAccessorsBroker;
+      public event global::System.EventHandler EventFromAccessors
+      {
+        add
+        {
+          global::Metalama.Framework.RunTime.ActionEventBroker<global::System.EventHandler, (global::System.Object? , global::System.EventArgs)>.EnsureInitialized(ref this._eventFromAccessorsBroker, this, EventFromAccessorsBrokerCallbacks_0);
+          this._eventFromAccessorsBroker.AddHandler(value);
+        }
+        remove
+        {
+          this._eventFromAccessorsBroker?.RemoveHandler(value);
+        }
+      }
+    }
