+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/Invokers/Events/BaseClassStatic_UserHidden.t.cs
@@ -1,6 +1,6 @@
 public class TargetClass : BaseClass
 {
-  public new static event EventHandler Event
+  public static new event EventHandler Event
   {
     add
     {
