+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/Invokers/Events/BaseClassStatic_UserHidden_AspectOverride.t.cs
@@ -1,7 +1,7 @@
 public class TargetClass : BaseClass
 {
   [OverrideAspect]
-  public new static event EventHandler Event
+  public static new event EventHandler Event
   {
     add
     { // Invoke TargetClass.Event_Source
