+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/Invokers/Events/TargetClass_DifferentInstance.cs
@@ -30,9 +30,9 @@ public class InvokerAspect : EventAspect
     [Template]
     public void AddTemplate( [CompileTime] IEvent target )
     {
-        meta.InsertComment( "Invoke instance.Event" );
+        meta.InsertComment( "Invoke _instance.Event" );
         target.With( (IExpression?)meta.Target.Event.DeclaringType.Fields.Single().Value! ).Add( meta.RunTime( TargetClass.StaticTarget ) );
-        meta.InsertComment( "Invoke instance.Event" );
+        meta.InsertComment( "Invoke _instance.Event" );
 
         target.With( (IExpression?)meta.Target.Event.DeclaringType.Fields.Single().Value!, InvokerOptions.Final )
             .Add( meta.RunTime( TargetClass.StaticTarget ) );
@@ -43,9 +43,9 @@ public class InvokerAspect : EventAspect
     [Template]
     public void RemoveTemplate( [CompileTime] IEvent target )
     {
-        meta.InsertComment( "Invoke instance.Event" );
+        meta.InsertComment( "Invoke _instance.Event" );
         target.With( (IExpression?)meta.Target.Event.DeclaringType.Fields.Single().Value! ).Remove( meta.RunTime( TargetClass.StaticTarget ) );
-        meta.InsertComment( "Invoke instance.Event" );
+        meta.InsertComment( "Invoke _instance.Event" );
 
         target.With( (IExpression?)meta.Target.Event.DeclaringType.Fields.Single().Value!, InvokerOptions.Final )
             .Remove( meta.RunTime( TargetClass.StaticTarget ) );
