+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/Invokers/Events/TargetClass_DifferentInstance.t.cs
@@ -9,21 +9,21 @@ public class TargetClass
     {
     }
   }
-  private TargetClass? instance;
+  private TargetClass? _instance;
   [InvokerAspect]
   public event EventHandler Invoker
   {
     add
-    { // Invoke instance.Event
-      this.instance!.Event += global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Invokers.Events.TargetClass_DifferentInstance.TargetClass.StaticTarget;
-      // Invoke instance.Event
-      this.instance!.Event += global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Invokers.Events.TargetClass_DifferentInstance.TargetClass.StaticTarget;
+    { // Invoke _instance.Event
+      this._instance!.Event += global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Invokers.Events.TargetClass_DifferentInstance.TargetClass.StaticTarget;
+      // Invoke _instance.Event
+      this._instance!.Event += global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Invokers.Events.TargetClass_DifferentInstance.TargetClass.StaticTarget;
     }
     remove
-    { // Invoke instance.Event
-      this.instance!.Event -= global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Invokers.Events.TargetClass_DifferentInstance.TargetClass.StaticTarget;
-      // Invoke instance.Event
-      this.instance!.Event -= global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Invokers.Events.TargetClass_DifferentInstance.TargetClass.StaticTarget;
+    { // Invoke _instance.Event
+      this._instance!.Event -= global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Invokers.Events.TargetClass_DifferentInstance.TargetClass.StaticTarget;
+      // Invoke _instance.Event
+      this._instance!.Event -= global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Invokers.Events.TargetClass_DifferentInstance.TargetClass.StaticTarget;
     }
   }
   public static void StaticTarget(object? sender, EventArgs args)
