+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/Invokers/Events/TargetClass_DifferentInstance_Current.t.cs
@@ -9,17 +9,17 @@ public class TargetClass
     {
     }
   }
-  private TargetClass? instance;
+  private TargetClass? _instance;
   [InvokerAspect]
   public event EventHandler Invoker
   {
     add
     {
-      this.instance!.Event += global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Invokers.Events.TargetClass_DifferentInstance_Current.TargetClass.StaticTarget;
+      this._instance!.Event += global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Invokers.Events.TargetClass_DifferentInstance_Current.TargetClass.StaticTarget;
     }
     remove
     {
-      this.instance!.Event -= global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Invokers.Events.TargetClass_DifferentInstance_Current.TargetClass.StaticTarget;
+      this._instance!.Event -= global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Invokers.Events.TargetClass_DifferentInstance_Current.TargetClass.StaticTarget;
     }
   }
   public static void StaticTarget(object? sender, EventArgs args)
