+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/Overrides/EventFields/AsMethod_Raise.t.cs
@@ -1,6 +1,6 @@
 internal class TargetClass
 {
-  private static readonly global::Metalama.Framework.RunTime.ActionEventBrokerCallbacks<global::System.EventHandler?, (global::System.Object? , global::System.EventArgs)> EventBrokerCallbacks_0 = new global::Metalama.Framework.RunTime.ActionEventBrokerCallbacks<global::System.EventHandler?, (global::System.Object? , global::System.EventArgs)>(static (handler, me, args) => ((global::Metalama.Framework.IntegrationTests.Aspects.Overrides.EventFields.AsMethod_Raise.TargetClass)me).Event_Raise_Override(handler, args), static b => (sender, e) => b.Invoke((sender, e)), static (handler, me) => ((global::Metalama.Framework.IntegrationTests.Aspects.Overrides.EventFields.AsMethod_Raise.TargetClass)me).Event_Override += handler, static (handler, me) => ((global::Metalama.Framework.IntegrationTests.Aspects.Overrides.EventFields.AsMethod_Raise.TargetClass)me).Event_Override -= handler);
+  private static readonly global::Metalama.Framework.RunTime.ActionEventBrokerCallbacks<global::System.EventHandler?, (global::System.Object? , global::System.EventArgs)> EventBrokerCallbacks_0 = new global::Metalama.Framework.RunTime.ActionEventBrokerCallbacks<global::System.EventHandler?, (global::System.Object? , global::System.EventArgs)>(static (handler, me, args) => ((global::Metalama.Framework.IntegrationTests.Aspects.Overrides.EventFields.AsMethod_Raise.TargetClass)me).Event_Invoke_Override(handler, args), static b => (sender, e) => b.Invoke((sender, e)), static (handler, me) => ((global::Metalama.Framework.IntegrationTests.Aspects.Overrides.EventFields.AsMethod_Raise.TargetClass)me).Event_Override += handler, static (handler, me) => ((global::Metalama.Framework.IntegrationTests.Aspects.Overrides.EventFields.AsMethod_Raise.TargetClass)me).Event_Override -= handler);
   private event EventHandler? _event;
   private volatile global::Metalama.Framework.RunTime.ActionEventBroker<global::System.EventHandler?, (global::System.Object? , global::System.EventArgs)>? _eventBroker;
   [Override]
@@ -27,7 +27,7 @@ internal class TargetClass
       this._event -= value;
     }
   }
-  private void Event_Raise_Override(global::System.EventHandler? handler, (global::System.Object? sender, global::System.EventArgs e) args)
+  private void Event_Invoke_Override(global::System.EventHandler? handler, (global::System.Object? sender, global::System.EventArgs e) args)
   {
     global::System.Console.WriteLine("Overridden raise");
     handler.Invoke(args.sender, args.e);
