+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/Overrides/EventFields/InvokeAddRemove.cs
@@ -11,18 +11,16 @@ using Metalama.Framework.Advising;
 using Metalama.Framework.Aspects;
 using Metalama.Framework.Code;
 
-namespace Metalama.Framework.IntegrationTests.Aspects.Introductions.Events.RaiseHandler
+namespace Metalama.Framework.Tests.AspectTests.Tests.Aspects.Overrides.EventFields.InvokeAddRemove
 {
-    public class IntroductionAttribute : TypeAspect
+    public class OverrideAttribute : EventAspect
     {
-        public override void BuildAspect( IAspectBuilder<INamedType> builder )
+        public override void BuildAspect( IAspectBuilder<IEvent> builder )
         {
-            builder.IntroduceEvent(
-                "EventFromAccessors",
+            builder.OverrideAccessors(
                 nameof( AddEventTemplate ),
                 nameof( RemoveEventTemplate ),
-                nameof( InvokeEventTemplate ),
-                buildEvent: e => e.Accessibility = Accessibility.Public );
+                invokeTemplate: nameof( InvokeEventTemplate ));
         }
 
         [Template]
@@ -48,6 +46,9 @@ namespace Metalama.Framework.IntegrationTests.Aspects.Introductions.Events.Raise
     }
 
     // <target>
-    [Introduction]
-    internal class TargetClass { }
+    internal class TargetClass 
+    {
+        [Override]
+        public event EventHandler Event;
+    }
 }
