+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/Overrides/Events/InvokeInvalidTemplate.t.cs
