+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/Overrides/Events/Invoke_CustomDelegate.cs
@@ -2,10 +2,6 @@
 // SharpCrafters s.r.o. licenses this file to you under either the MIT license or a proprietary license, depending on the repository from which it was obtained.
 // Refer to LICENSE.md in the repository root for complete details.
 
-#if TESTOPTIONS
-// @RequiredConstant(NET6_0_OR_GREATER)
-#endif
-
 using System;
 using Metalama.Framework.Advising;
 using Metalama.Framework.Aspects;
@@ -13,41 +9,41 @@ using Metalama.Framework.Code;
 
 #pragma warning disable IDE0052
 
-namespace Metalama.Framework.IntegrationTests.Aspects.Overrides.Events.AsMethod_Raise;
+namespace Metalama.Framework.Tests.AspectTests.Tests.Aspects.Overrides.Events.Invoke_CustomDelegate;
+
+public delegate void MyEventHandler( object sender, int args1, int arg2 );
 
 public class OverrideAttribute : EventAspect
 {
     public override void BuildAspect( IAspectBuilder<IEvent> builder )
     {
-        builder.With( builder.Target.RaiseMethod ).Override( nameof( InvokeEventTemplate ) );
+        builder.OverrideAccessors( invokeTemplate: nameof( InvokeEventTemplate ));
     }
 
     [Template]
-    public void InvokeEventTemplate( EventHandler value )
+    public void InvokeEventTemplate()
     {
-        Console.WriteLine( "Overridden invoke" );
+        Console.WriteLine( "Invoke" );
         meta.Proceed();
     }
 }
 
 // <target>
-internal class TargetClass
+internal class TargetClass 
 {
-    private EventHandler? _handler;
+    private MyEventHandler? _handler;
 
     [Override]
-    public event EventHandler Event
+    public event MyEventHandler Event
     {
         add
         {
             this._handler = value;
-            Console.WriteLine( "OriginalAdd" );
         }
 
         remove
         {
             this._handler = null;
-            Console.WriteLine( "OriginalRemove" );
         }
     }
 }
