+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/Overrides/Events/Invoke_Double.cs
@@ -13,19 +13,34 @@ using Metalama.Framework.Code;
 
 #pragma warning disable IDE0052
 
-namespace Metalama.Framework.IntegrationTests.Aspects.Overrides.Events.Raise;
+namespace Metalama.Framework.Tests.AspectTests.Tests.Aspects.Overrides.Events.Invoke_Double;
 
 public class OverrideAttribute : EventAspect
 {
     public override void BuildAspect( IAspectBuilder<IEvent> builder )
     {
-        builder.OverrideAccessors( null, null, nameof( RaiseEventTemplate ));
+        builder.OverrideAccessors( nameof( AddEventTemplate ), nameof( RemoveEventTemplate ), invokeTemplate: nameof( InvokeEventTemplate ), args: new { ordinal = 0 } );
+        builder.OverrideAccessors( nameof( AddEventTemplate ), nameof( RemoveEventTemplate ), invokeTemplate: nameof( InvokeEventTemplate ), args: new { ordinal = 1 } );
     }
 
     [Template]
-    public void RaiseEventTemplate()
+    public void AddEventTemplate( [CompileTime] int ordinal )
     {
-        Console.WriteLine( "Raise" );
+        Console.WriteLine( $"Add {ordinal}" );
+        meta.Proceed();
+    }
+
+    [Template]
+    public void RemoveEventTemplate( [CompileTime] int ordinal )
+    {
+        Console.WriteLine( $"Remove {ordinal}" );
+        meta.Proceed();
+    }
+
+    [Template]
+    public void InvokeEventTemplate( [CompileTime] int ordinal )
+    {
+        Console.WriteLine( $"Invoke {ordinal}" );
         meta.Proceed();
     }
 }
