+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/Overrides/Events/Invoke_Double.t.cs
@@ -1,0 +1,69 @@
+internal class TargetClass
+{
+  private static readonly global::Metalama.Framework.RunTime.ActionEventBrokerCallbacks<global::System.EventHandler, (global::System.Object? , global::System.EventArgs)> EventBrokerCallbacks_0 = new global::Metalama.Framework.RunTime.ActionEventBrokerCallbacks<global::System.EventHandler, (global::System.Object? , global::System.EventArgs)>(static (handler, me, args) => ((global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Overrides.Events.Invoke_Double.TargetClass)me).Event_Invoke_Override(handler, args), static b => (sender, e) => b.Invoke((sender, e)), static (handler, me) => ((global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Overrides.Events.Invoke_Double.TargetClass)me).Event_Override += handler, static (handler, me) => ((global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Overrides.Events.Invoke_Double.TargetClass)me).Event_Override -= handler);
+  private static readonly global::Metalama.Framework.RunTime.ActionEventBrokerCallbacks<global::System.EventHandler, (global::System.Object? , global::System.EventArgs)> EventBrokerCallbacks_2 = new global::Metalama.Framework.RunTime.ActionEventBrokerCallbacks<global::System.EventHandler, (global::System.Object? , global::System.EventArgs)>(static (handler, me, args) => ((global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Overrides.Events.Invoke_Double.TargetClass)me).Event_Invoke_Override1(handler, args), static b => (sender, e) => b.Invoke((sender, e)), static (handler, me) => ((global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Overrides.Events.Invoke_Double.TargetClass)me).Event_Override1 += handler, static (handler, me) => ((global::Metalama.Framework.Tests.AspectTests.Tests.Aspects.Overrides.Events.Invoke_Double.TargetClass)me).Event_Override1 -= handler);
+  private EventHandler? _handler;
+  private volatile global::Metalama.Framework.RunTime.ActionEventBroker<global::System.EventHandler, (global::System.Object? , global::System.EventArgs)>? _eventBroker;
+  private volatile global::Metalama.Framework.RunTime.ActionEventBroker<global::System.EventHandler, (global::System.Object? , global::System.EventArgs)>? _eventBroker_1;
+  [Override]
+  public event EventHandler Event
+  {
+    add
+    {
+      global::Metalama.Framework.RunTime.ActionEventBroker<global::System.EventHandler, (global::System.Object? , global::System.EventArgs)>.EnsureInitialized(ref this._eventBroker_1, this, EventBrokerCallbacks_2);
+      this._eventBroker_1.AddHandler(value);
+    }
+    remove
+    {
+      this._eventBroker_1?.RemoveHandler(value);
+    }
+  }
+  private event global::System.EventHandler Event_Override
+  {
+    add
+    {
+      global::System.Console.WriteLine("Add 0");
+      this._handler = value;
+    }
+    remove
+    {
+      global::System.Console.WriteLine("Remove 0");
+      this._handler = null;
+    }
+  }
+  public event global::System.EventHandler Event_OverrideBrokered
+  {
+    add
+    {
+      global::Metalama.Framework.RunTime.ActionEventBroker<global::System.EventHandler, (global::System.Object? , global::System.EventArgs)>.EnsureInitialized(ref this._eventBroker, this, EventBrokerCallbacks_0);
+      this._eventBroker.AddHandler(value);
+    }
+    remove
+    {
+      this._eventBroker?.RemoveHandler(value);
+    }
+  }
+  private void Event_Invoke_Override(global::System.EventHandler handler, (global::System.Object? sender, global::System.EventArgs e) args)
+  {
+    global::System.Console.WriteLine("Invoke 0");
+    handler.Invoke(args.sender, args.e);
+  }
+  private event global::System.EventHandler Event_Override1
+  {
+    add
+    {
+      global::System.Console.WriteLine("Add 1");
+      this.Event_OverrideBrokered += value;
+    }
+    remove
+    {
+      global::System.Console.WriteLine("Remove 1");
+      this.Event_OverrideBrokered -= value;
+    }
+  }
+  private void Event_Invoke_Override1(global::System.EventHandler handler, (global::System.Object? sender, global::System.EventArgs e) args)
+  {
+    global::System.Console.WriteLine("Invoke 1");
+    handler.Invoke(args.sender, args.e);
+  }
+}
