--- Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/Overrides/Events/RaiseInvalidDelegate.t.cs
@@ -1,3 +1,0 @@
-// CompileTimeAspectPipeline.ExecuteAsync failed.
-// Error LAMA0041 on `EventWithReturn`: `'Exception of type 'System.InvalidOperationException' thrown while executing BuildAspect for aspect [OverrideAttribute] applied to 'TargetClass.EventWithReturn': Cannot add an OverrideEventRaise advice to 'TargetClass.EventWithReturn' because ''TargetClass.EventWithReturn'' must have delegate type with void return value. Check the IsAdviceEligible(AdviceKind.OverrideEventRaise) method. Exception details are in '(none)'. To attach a debugger to the compiler, use the  '-p:MetalamaDebugCompiler=True' command-line option.`
-// Error LAMA0041 on `EventWithRef`: `'Exception of type 'System.InvalidOperationException' thrown while executing BuildAspect for aspect [OverrideAttribute] applied to 'TargetClass.EventWithRef': Cannot add an OverrideEventRaise advice to 'TargetClass.EventWithRef' because ''TargetClass.EventWithRef'' must have delegate type without a parameter of out/ref/in/pointer type. Check the IsAdviceEligible(AdviceKind.OverrideEventRaise) method. Exception details are in '(none)'. To attach a debugger to the compiler, use the  '-p:MetalamaDebugCompiler=True' command-line option.`
