+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/Samples/SafeEvent.t.cs
@@ -1,7 +1,7 @@
 internal class TargetCode
 {
-  private static readonly ActionEventBrokerCallbacks<EventHandler, (object? , EventArgs)> EventFieldBrokerCallbacks_0 = new ActionEventBrokerCallbacks<EventHandler, (object? , EventArgs)>(static (handler, me, args) => ((TargetCode)me).EventField_Raise_SafeEvent(handler, args), static b => (sender, e) => b.Invoke((sender, e)), static (handler, me) => ((TargetCode)me).EventField_SafeEvent += handler, static (handler, me) => ((TargetCode)me).EventField_SafeEvent -= handler);
-  private static readonly ActionEventBrokerCallbacks<EventHandler, (object? , EventArgs)> EventBrokerCallbacks_0 = new ActionEventBrokerCallbacks<EventHandler, (object? , EventArgs)>(static (handler, me, args) => ((TargetCode)me).Event_Raise_SafeEvent(handler, args), static b => (sender, e) => b.Invoke((sender, e)), static (handler, me) => ((TargetCode)me).Event_SafeEvent += handler, static (handler, me) => ((TargetCode)me).Event_SafeEvent -= handler);
+  private static readonly ActionEventBrokerCallbacks<EventHandler, (object? , EventArgs)> EventFieldBrokerCallbacks_0 = new ActionEventBrokerCallbacks<EventHandler, (object? , EventArgs)>(static (handler, me, args) => ((TargetCode)me).EventField_Invoke_SafeEvent(handler, args), static b => (sender, e) => b.Invoke((sender, e)), static (handler, me) => ((TargetCode)me).EventField_SafeEvent += handler, static (handler, me) => ((TargetCode)me).EventField_SafeEvent -= handler);
+  private static readonly ActionEventBrokerCallbacks<EventHandler, (object? , EventArgs)> EventBrokerCallbacks_0 = new ActionEventBrokerCallbacks<EventHandler, (object? , EventArgs)>(static (handler, me, args) => ((TargetCode)me).Event_Invoke_SafeEvent(handler, args), static b => (sender, e) => b.Invoke((sender, e)), static (handler, me) => ((TargetCode)me).Event_SafeEvent += handler, static (handler, me) => ((TargetCode)me).Event_SafeEvent -= handler);
   private List<EventHandler> _delegates = new List<EventHandler>();
   private event EventHandler _eventField = default !;
   private volatile ActionEventBroker<EventHandler, (object? , EventArgs)>? _eventFieldBroker;
@@ -29,7 +29,7 @@ internal class TargetCode
       this._eventField -= value;
     }
   }
-  private void EventField_Raise_SafeEvent(EventHandler handler, (object? sender, EventArgs e) args)
+  private void EventField_Invoke_SafeEvent(EventHandler handler, (object? sender, EventArgs e) args)
   {
     try
     {
@@ -79,7 +79,7 @@ internal class TargetCode
       this.Event_Source -= value;
     }
   }
-  private void Event_Raise_SafeEvent(EventHandler handler, (object? sender, EventArgs e) args)
+  private void Event_Invoke_SafeEvent(EventHandler handler, (object? sender, EventArgs e) args)
   {
     try
     {
